+++ src/exact-key-map/index.ts
@@ -1,3 +1,1 @@
 export * from './ExactKeyMap';
-export * from './LooseExactKeyMap';
-export * from './ConstExactKeyMap';