+++ src/utils/index.ts
@@ -1,5 +1,3 @@
 export * from './isArrayOfTuples';
-export * from './isEntriesArray';
 export * from './isPlainObject';
 export * from './isTuple';
-export * from './toEntriesArray';